test: GroupMembershipService 테스트 추가
- 기존 GroupService 테스트에서 분리
